implement student module, update  admin dashboard templates and controller

+++ src/main/java/com/edu/miusched/controller/CourseController.java
@@ -1,19 +1,19 @@
 package com.edu.miusched.controller;
 
 
+import com.edu.miusched.AjaxResponse;
 import com.edu.miusched.domain.Course;
 import com.edu.miusched.domain.Entry;
 import com.edu.miusched.service.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,8 +22,10 @@ import java.util.List;
 public class CourseController {
   @Autowired
     CourseService courseService;
+    AjaxResponse response = new AjaxResponse();
     @RequestMapping(value = "/admin/courses", method = RequestMethod.GET)
     public String ListCourses(Model model) {
+
         model.addAttribute("courses", courseService.findAll());
 
         return "Admin/ManageCourse";
@@ -32,9 +34,13 @@ public class CourseController {
 
     @RequestMapping(value = {"/admin/course"}, method = RequestMethod.GET)
     public String getCourseHome(@ModelAttribute("newCourse") Course course, Model model) {
+       // AjaxResponse response = new AjaxResponse();
+        //response.success =false ?false :true;
+        System.out.println(response.success);
         Course  course1 = new Course();
         List<Course> courses = new ArrayList<Course>();
         courses.addAll(courseService.findAll());
+        model.addAttribute("response",response);
         model.addAttribute("courses", courses);
         model.addAttribute("newCourse", course);
         model.addAttribute("course1",course1);
@@ -84,9 +90,31 @@ public class CourseController {
         return "redirect:/admin/course";
     }
 
-    @RequestMapping(value="/admin/course/delete/{id}", method=RequestMethod.GET)
-    public String deleteCourse(@PathVariable("id") Long id) {
-        courseService.deleteCourse(id);
+//    @RequestMapping(value="/admin/course/delete/{id}", method=RequestMethod.GET)
+//    public String deleteCourse(@PathVariable("id") Long id) {
+//        courseService.deleteCourse(id);
+//        return "redirect:/admin/course";
+//    }
+
+    @RequestMapping(value = "/admin/course/delete/{id}", method = RequestMethod.GET)
+    public String delete(@PathVariable(value = "id") Long id,Model model) {
+        Course course = courseService.findCoursebyId(id);
+
+        if (course != null) {
+            try {
+                courseService.deleteCourse(id);
+                response.success = true;
+                response.msg = "Successfully deleted.";
+            } catch (DataIntegrityViolationException ignore) {
+                // Cannot remove course that is prerequisite of other course.
+                response.success = false;
+                response.msg = "Cannot remove course that is prerequisite of other course.";
+            }
+
+
+        }
+
+        model.addAttribute("response",response);
         return "redirect:/admin/course";
     }
